Adjusted component structure in order to avoid unnecessary rerenders
